\name{MixABEL-package}
\alias{MixABEL-package}
\title{MixABEL package...}
\description{MixABEL package}
\details{Package for fast (genome-wide) association analysis 
using mixed models. Main functions: 
\link{FastMixedModel}, \link{GWFGLS}}
\author{Yurii Aulchenko, William Astle, Erik Roos, Marcel Kempenaar}
