\name{summaryFGLS}
\alias{summaryFGLS}
\title{summary for FGLS...}
\usage{summaryFGLS(x, verbosity=1, varcov=FALSE, include.means=FALSE)}
\description{summary for FGLS}
\details{summary for Feasible GLS}
\value{one-line summary of output from FGLS function}
\author{Yurii Aulchenko}
\arguments{\item{x}{object returned by FGLS function}
\item{verbosity}{what to report; 0 -- only test stats, 
1 -- test stats and estimates concerning 'whichtest' parameters, 
2 -- test stats and estimates of all parameters}
\item{varcov}{wheather var-cov matrix for estimated parameters 
is to be reported}
\item{include.means}{weather mean values of predictors are to 
be reported}}
